rename base package to station
