+++ src/main/java/com/codecool/spaceship/model/station/SpaceStation.java
@@ -1,4 +1,4 @@
-package com.codecool.spaceship.model.base;
+package com.codecool.spaceship.model.station;
 
 import com.codecool.spaceship.model.Resource;
 import com.codecool.spaceship.model.Upgradeable;
@@ -12,7 +12,6 @@ import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 public class SpaceStation {
     private String name;
