+++ src/test/java/com/codecool/spaceship/model/station/SpaceStationStorageTest.java
@@ -1,4 +1,4 @@
-package com.codecool.spaceship.model.base;
+package com.codecool.spaceship.model.station;
 
 import com.codecool.spaceship.model.Resource;
 import com.codecool.spaceship.model.exception.StorageException;
